+++ lib/aua-mite/agents/standard_mite_client.rb
@@ -1,6 +1,6 @@
 module Aua::Agents::StandardMiteClient
   
-  KNOWN_CLIENTS = %w(mite.php mite.desk mite-rb jmite Redmine2mite Mantis2mite Billomat hearttp centralstationcrm)
+  KNOWN_CLIENTS = %w(git2mite Mite-on-rails mite.php mite.desk mite-rb jmite Redmine2mite Mantis2mite Billomat hearttp centralstationcrm)
   
   def self.extend?(agent)
     KNOWN_CLIENTS.include?(agent.app)