+++ CHANGES.md
@@ -2,6 +2,7 @@
 
 [full changelog](http://github.com/yolk/aua-mite/compare/v0.0.1...master)
 
+* Added support for Mite-on-rails and git2mite
 * Fixed specs
 
 ### 0.0.1 / 2011-01-27